Add updater, setup script, & readme
