+++ .env.example
@@ -1,5 +1,37 @@
+####
+##                                       .ENV INFO
+##
+#  The .env file is an easily-accessible configuration file that allows you to specify certain
+#  default behaviors for the bot. Any modifications to this file will require a restart to take
+#  effect. You can restart Discord TTL with the following command:
+##
+#                                  docker-compose restart
+##
+####
+
+####
+##                               Variable: DISCORD_BOT_TOKEN
+##
+#  This is an access token that will allow Discord TTL to run on a Discord bot of your choice.
+#  See https://ayu.dev/r/discord-bot-token-guide to learn how to generate a Discord bot token.
+##
+####
 DISCORD_BOT_TOKEN=
-# Number of seconds for Discord messages to live until.
-# Keep in mind that TTL may not be upheld if Discord is ratelimiting the bot.
-# 604800 seconds = 7 days
-MESSAGE_TTL=604800
+
+####
+##                              Variable: DEFAULT_MESSAGE_TTL
+##
+#  This field specifies the default number of seconds for Discord messages to live until. Keep 
+#  in mind that TTL may not be upheld if Discord is ratelimiting the bot. Note: TTL can be
+#  customized server-by-server with commands; This default simply declares what the configuration 
+#  will initially be set as before any commands are used.
+##
+#  Uncomment DEFAULT_MESSAGE_TTL (by removing the # in front of DEFAULT_MESSAGE_TTL below) if 
+#  you would like for a default TTL to be applied to all servers that add this TTL instance.
+##
+#  Comment DEFAULT_MESSAGE_TTL out (by adding a # in front of DEFAULT_MESSAGE_TTL) if you want 
+#  the default TTL to be unset, which will prevent TTL from deleting any messages until a TTL is 
+#  set via server commands.
+##
+####
+DEFAULT_MESSAGE_TTL=604800
