+++ .gitignore
@@ -3,3 +3,4 @@
 .env
 yarn-error.log
 *.db
+/data/
