+++ README.md
@@ -9,15 +9,6 @@ anything you wont expect. It also allows you to have ratelimits If you would lik
 
 # Steps to self-host
 
-## If you know what you're doing
-To get started, clone this repository and install [Docker](https://docs.docker.com/engine/install/ubuntu/) 
-& [Docker Compose](https://docs.docker.com/compose/install/). Then, populate a new `.env` file with a `DISCORD_BOT_TOKEN` 
-from the [Discord Dev Portal](https://discord.com/developers/applications). You can find an example of this in the 
-[`.env.example` on GitHub](https://github.com/ayubun/discord-ttl/blob/main/.env.example). Finally, you can start Discord TTL with Docker Compose:
-```
-docker-compose up -d
-```
-
 ## If you are new to self-hosting
 Firsly, I am happy that you are looking to expand your knowledge into this domain! To get started, you'll want a server to host your bot on.
 If you already have one, that's great! For my own sanity, this tutorial will only cover the steps to setting up an Ubuntu
@@ -27,21 +18,22 @@ server, but it is possible to get Discord TTL running on any OS, so long as you 
 If you don't have a VPS yet, you can find them all over the internet :D If you need a recommendation, GalaxyGate's Standard 1GB offers a 
 perfectly-capable $3/month option that might be within budget for most: https://galaxygate.net/hosting/vps/
 
-Once you have a VPS, clone the repository to a desired location (such as your home directory):
-```
+## Once you have a suitable location to host the bot
+Clone the repository to a desired location (such as your home directory):
+```bash
 cd ~ && git clone https://github.com/ayubun/discord-ttl
 ```
 Then, navigate to the newly-created `discord-ttl` directory:
-```
+```bash
 cd discord-ttl
 ```
-then, run the `setup.sh` script and follow any instructions:
-```
+Lastly, run the `setup.sh` script. Follow any instructions provided by the script:
+```bash
 ./setup.sh
 ```
 And you're done! :)
 
 **Note**: You will need a Discord bot token from the Discord Developers portal to be able to run Discord TTL. Discord.js has a really
-helpful guide on how to get one here: https://discordjs.guide/preparations/setting-up-a-bot-application.html
+helpful guide on how to get one here: https://ayu.dev/r/discord-bot-token-guide
 
-Once you get one, just paste it into the `.env` file (NOT the `.env.example`!) where it says `DISCORD_BOT_TOKEN=`
+Once you get a bot token, paste it into the `.env` file where it says `DISCORD_BOT_TOKEN=`
