+++ package.json
@@ -4,6 +4,7 @@
   "main": "dist/app.js",
   "repository": "git@github.com:ayubun/discord-ttl.git",
   "author": "Ayu <ayu@ayu.dev>",
+  "license": "MIT",
   "precommit": [
     "yarn run lint",
     "yarn run pretty"
@@ -12,10 +13,12 @@
     "plugin:you-dont-need-momentjs/recommended"
   ],
   "dependencies": {
+    "@types/sqlite3": "^3.1.8",
     "discord.js": "^13.8.1",
     "dotenv": "^16.0.0",
     "eslint": "^8.2.0",
     "pg": "^8.7.1",
+    "sqlite3": "^5.1.6",
     "ts-node": "^10.4.0",
     "typescript": "^4.4.4"
   },
